Feat/deactivate privileges when state not live (#929)
We have identified an edge case where a user may move to a state that is
not 'live' in the system, but has uploaded license data into the system.
We don't practitioners to be able to move privileges over to the state
if it is not live. In this case, the privileges should be deactivated as
if there were no license for them in that state.

Closes #906
