+++ backend/compact-connect/tests/smoke/home_jurisdiction_change_smoke_tests.py
@@ -3,6 +3,7 @@ import json
 import time
 
 import requests
+from compact_configuration_smoke_tests import test_jurisdiction_configuration
 from config import config, logger
 from purchasing_privileges_smoke_tests import test_purchasing_privilege
 from smoke_common import (
@@ -41,6 +42,15 @@ def test_home_jurisdiction_change_inactivates_privileges_when_no_license_in_new_
 
     original_jurisdiction = provider_info_before.get('currentHomeJurisdiction')
     logger.info(f'Original home jurisdiction: {original_jurisdiction}')
+    new_jurisdiction = 'al'  # Alabama - assuming the provider doesn't have a license here
+
+
+    # we must ensure we have a valid live jurisdiction configuration in place for the current, new, and privilege
+    # states so the privilege can be moved over successfully
+    test_jurisdiction_configuration(jurisdiction=original_jurisdiction, recreate_compact_config=True)
+    test_jurisdiction_configuration(jurisdiction=new_jurisdiction, recreate_compact_config=False)
+    # privilege jurisdiction
+    test_jurisdiction_configuration(jurisdiction='ne', recreate_compact_config=False)
 
     # Purchase a privilege for the provider
     # This uses the same test_purchasing_privilege function from purchasing_privileges_smoke_tests.py
@@ -58,8 +68,6 @@ def test_home_jurisdiction_change_inactivates_privileges_when_no_license_in_new_
         )
 
     # Now change the home jurisdiction to one where the provider doesn't have a license
-    new_jurisdiction = 'al'  # Alabama - assuming the provider doesn't have a license here
-
     logger.info(f'Changing home jurisdiction to {new_jurisdiction}')
 
     response = requests.put(
@@ -212,7 +220,6 @@ def add_license_for_provider(provider_record: dict, jurisdiction: str):
     time.sleep(1)
 
     return license_record
-
 
 def test_home_jurisdiction_change_moves_privileges_when_valid_license_in_new_jurisdiction():
     """
